Data - UK Butterfly Monitoring Scheme - Wider Countryside Species

This dataset from the UK Butterfly Monitoring Scheme has national abundance indices for 24 species regarded as wider countryside species from 1976-2017. There are 1005 lines of data, reflecting late entry into the time series for the Scotch Argus (Erebia aethops). Time-series for the other 23 species are complete.

Usage

data(butterflies_wc)

Format

There are four columns of data:

"species" The species' name (Latin binomial)

"year" The year to which the index value refers (integer)

"collated_index" A "collated" index of abundance on the log10 scale. The time-series for each species is fixed to have a mean of exactly 2. Since 10^2 = 100, the time-series are centered on a value of 100.

"se" standard errors on the index value (on the log10 scale), as estimated by the underlying statistical model.

Source

https://catalogue.ceh.ac.uk/documents/ace3c3ef-df89-40b9-ba8b-106997fd6d9c

References

Botham, M.S.; Brereton, T.; Harris, S.; Harrower, C.; Middlebrook, I.; Randle, Z.; Roy, D.B. (2019). United Kingdom Butterfly Monitoring Scheme: collated indices 2017. NERC Environmental Information Data Centre. https://catalogue.ceh.ac.uk/documents/ace3c3ef-df89-40b9-ba8b-106997fd6d9c

Examples

data(butterflies_wc)
head(butterflies_wc)
#>          species year    index         se
#> 1 Aglais urticae 1976 2.105378 0.01652560
#> 2 Aglais urticae 1977 1.972035 0.01565523
#> 3 Aglais urticae 1978 2.277109 0.01259163
#> 4 Aglais urticae 1979 2.115708 0.01260353
#> 5 Aglais urticae 1980 2.135539 0.01243311
#> 6 Aglais urticae 1981 1.892195 0.01375867