Rescale species values for indicator

This function takes in a dataframe of multi-species data and rescales them so that the value in the starting year is the same. This function accounts for species that have no data at the beginning of the period or those that have no data at teh end of the period.

Usage

rescale_species(Data, index = 100, max = 10000, min = 1)

Arguments

Data

A matrix, the first column, named year, gives the year. Subsequent named columns give the species. Values in the table are the yearly values to be rescaled.

index

The index value for the first year, defaults to 100.

max

The upper limit allowed for scaled values. Values greater than this will be set equal to max.

min

The upper limit allowed for scaled values. Values greater than this will be set equal to min.

Value

A matrix. In each row we have the year, the species and the scaled value. There is also an additional column, 'geomean' giving the geometric mean for each year.

Examples

# We will create fake data which will mimic a summarised output from a sparta package occupancy detection model. Please see the vignette for more details.

# Create a matrix with random data using set.seed for reproducibility
set.seed(123)
data <- matrix(rnorm(50 * 27), nrow = 50, ncol = 27)

# Assign column names
col_names <- c("year", letters[1:26])
dimnames(data) <- list(NULL, col_names)

# Create the 'trends_summary' object
trends_summary <- data

# Re-scale the data so that the value for all species in the first year is the same (all defaults used).
# This function accounts for species that have no data at the beginning of the dataset by entering them at the geometric mean for that year, this stops them dramatically changing the indicator value in the year they join the dataset.
# It also accounts for species that leave the dataset before the end by holding them at their last value.
# Finally it limits the species values that can be given, preventing extremely high or low values biasing the indicator.
rescale_species(Data = trends_summary)
#>              year         a          b          c          d         e
#>  [1,] -0.56047565 100.00000 100.000000 100.000000 100.000000 100.00000
#>  [2,] -0.23017749   1.00000   1.000000  97.626548  59.687411 149.59320
#>  [3,]  1.55870831   1.00000  34.725450  42.171664   1.000000  91.56406
#>  [4,]  0.07050839 540.26935  48.921648   1.000000  24.703998   1.00000
#>  [5,]  0.12928774   1.00000 133.954079   1.000000   1.000000   1.00000
#>  [6,]  1.71506499 598.64184   6.338304   1.000000   1.000000  23.57946
#>  [7,]  0.46091621   1.00000 110.486658  71.469058   1.000000   1.00000
#>  [8,] -1.26506123 230.78209 234.786955   1.000000   1.000000   1.00000
#>  [9,] -0.68685285  48.89269  53.522383 124.022497  75.081849  30.25533
#> [10,] -0.44566197  85.24508   1.000000   1.000000   1.000000 408.11775
#> [11,]  1.22408180 149.86646  80.988407 133.637115   5.423171 138.74157
#> [12,]  0.35981383   1.00000   1.000000   1.000000  11.082694 130.42245
#> [13,]  0.40077145   1.00000 227.740394   1.000000  56.051941   1.00000
#> [14,]  0.11068272   1.00000   7.821150 411.435839   1.000000   1.00000
#> [15,] -0.55584113   1.00000   1.000000   1.000000   1.000000   1.00000
#> [16,]  1.78691314 119.82095   1.000000  37.858688  76.209253   1.00000
#> [17,]  0.49785048 176.93526   1.000000  80.809735   1.000000  35.44993
#> [18,] -1.96661716  20.92395  90.188639   1.000000   1.000000 467.65548
#> [19,]  0.70135590 364.07425 119.608178  65.613375   1.000000 103.50822
#> [20,] -0.47279141 809.29130 144.160941  46.838432   1.000000   1.00000
#> [21,] -1.06782371   1.00000   1.000000   1.000000   1.000000   1.00000
#> [22,] -0.21797491   1.00000 133.370757   8.288665  28.105462   1.00000
#> [23,] -1.02600445 397.02527  69.053056   1.000000  50.474937   1.00000
#> [24,] -0.72889123   1.00000  36.048680 270.197656  32.180509 371.47595
#> [25,] -0.62503927   1.00000   1.000000   1.000000   1.000000   1.00000
#> [26,] -1.68669331 404.85449  91.771379   1.000000   2.717376 118.89079
#> [27,]  0.83778704   1.00000   1.000000   4.797072   1.000000   1.00000
#> [28,]  0.15337312   1.00000   1.000000  39.414173   1.000000   1.00000
#> [29,] -1.13813694  71.57135 135.395235  55.414746  40.233142   1.00000
#> [30,]  1.25381492   1.00000  10.037645   1.000000   1.000000   1.00000
#> [31,]  0.42646422   2.27547   1.000000   1.000000  88.925085 443.94631
#> [32,] -0.29507148 152.09327   1.000000 160.355831   1.000000   1.00000
#> [33,]  0.89512566   1.00000   1.000000   1.000000   9.757041   1.00000
#> [34,]  0.87813349 254.37404  59.472541   1.000000   1.000000  70.72673
#> [35,]  0.82158108   1.00000 289.024247   1.000000   1.000000   1.00000
#> [36,]  0.68864025 130.97423   1.000000   1.000000   1.000000   1.00000
#> [37,]  0.55391765 432.98810 205.606218 140.899524   1.000000   1.00000
#> [38,] -0.06191171 171.79222   1.000000  10.757028  19.054959   1.00000
#> [39,] -0.30596266   1.00000   1.000000  95.723829  14.749082  58.31968
#> [40,] -0.38047100 453.50322 203.248849   1.000000   1.000000   1.00000
#> [41,] -0.69470698 392.19552   1.000000  27.222894   1.000000   1.00000
#> [42,] -0.20791728 216.48515  36.908089   1.000000   1.000000   1.00000
#> [43,] -1.26539635  94.24172 221.302032  12.006965  68.039550   1.00000
#> [44,]  2.16895597   1.00000 213.211382   1.000000   1.000000 153.74430
#> [45,]  1.20796200 537.13107 225.439383   1.000000   1.000000   1.00000
#> [46,] -1.12310858   1.00000  74.732773 253.537501  86.517776   1.00000
#> [47,] -0.40288484 863.47143 205.763243  76.257357   1.000000   1.00000
#> [48,] -0.46665535 605.01327   1.000000   1.000000   1.000000 359.66250
#> [49,]  0.77996512   1.00000   1.000000   1.000000   1.000000   1.00000
#> [50,] -0.08336907   1.00000 181.168157   1.000000  22.078302   1.00000
#>                f          g          h          i         j          k
#>  [1,] 100.000000 100.000000  100.00000 100.000000 100.00000 100.000000
#>  [2,] 105.235539   1.000000 1588.79102  73.170241 165.09560   1.000000
#>  [3,] 131.219707   1.000000  862.94537  30.431225   1.00000 131.460036
#>  [4,] 147.154810  11.492001   39.21032  49.998409   1.00000  29.568406
#>  [5,]  61.120491   1.000000    1.00000  64.120065 250.73675 195.515102
#>  [6,]   1.000000  32.898684 2243.93131   1.000000  15.80804   1.000000
#>  [7,] 281.612373  40.537237  475.49370  77.619916 148.85504   1.000000
#>  [8,]   1.000000   1.000000    1.00000   1.000000 344.03982   1.000000
#>  [9,]   1.000000   1.000000  732.51539  16.122515   1.00000   1.000000
#> [10,]   1.000000 253.359815    1.00000   1.000000  13.16043   1.000000
#> [11,]   1.000000   1.000000    1.00000  60.959419  16.17718   1.000000
#> [12,]   1.000000  64.160566    1.00000   1.000000   1.00000  68.495617
#> [13,] 241.418981  26.973578    1.00000  36.248806   1.00000   1.000000
#> [14,]  84.098326 100.958680  166.82341   1.000000   1.00000 175.644867
#> [15,]  49.220594  80.562091  562.39655   1.000000 102.41621   1.000000
#> [16,]   1.000000   1.000000 3593.38227  84.592319 122.08141  11.963385
#> [17,]  14.774203  37.259995  126.35406  51.796616  21.89805  63.082804
#> [18,] 175.975166   1.000000    1.00000 120.543863   1.00000  69.444004
#> [19,]   1.000000  84.430639    1.00000   4.554937 172.73512   1.000000
#> [20,]   1.000000   1.000000  754.90539  78.649392  30.62154   9.409586
#> [21,]   1.000000 238.119081    1.00000 138.102336   1.00000 113.771964
#> [22,]   1.000000   1.000000    1.00000   1.000000  17.98993   1.000000
#> [23,] 187.177758   1.000000    1.00000   1.000000 116.03738   1.000000
#> [24,]   1.000000  81.322766    1.00000   1.000000  45.84623  37.921731
#> [25,]  73.109834  50.262178  976.76061   1.000000   1.00000   1.000000
#> [26,]   1.000000   1.000000  612.23898  10.604045   1.00000 162.149026
#> [27,]   8.503687   1.000000    1.00000 119.707926   1.00000  79.684044
#> [28,]   1.000000  14.234857    1.00000   1.000000   1.00000   8.541412
#> [29,]   1.000000   1.000000    1.00000  26.076905   1.00000   1.000000
#> [30,]   1.000000   1.000000 1955.66136   1.000000  92.78344   3.089134
#> [31,]   1.000000   3.404237  259.00913   1.429717   1.00000   1.000000
#> [32,] 166.158269  18.742952    1.00000  21.955897  84.12354   1.000000
#> [33,] 100.889524  17.215387    1.00000  92.856021 236.01635   1.000000
#> [34,]   1.000000   1.000000 1367.17059   8.481416   1.00000  59.019475
#> [35,]   1.000000  46.912309    1.00000  49.835092   1.00000   1.000000
#> [36,] 286.926982 135.827323 1464.76205  54.450416   1.00000   1.000000
#> [37,] 190.709871  44.951916    1.00000  63.952165   1.00000  45.144267
#> [38,]  28.071752   1.000000    1.00000   1.000000   1.00000   1.000000
#> [39,]   1.000000   1.000000 3024.88868 113.735917 101.10723  69.000885
#> [40,]  14.244581  34.100788    1.00000   1.000000  33.60081   1.000000
#> [41,]   1.000000   1.000000 1687.20998   1.000000  45.39813   1.000000
#> [42,]   1.000000   1.000000    1.00000  98.157716  77.87097  55.032034
#> [43,]   1.000000  87.123185    1.00000  90.469924   1.00000  99.483443
#> [44,]   1.000000   1.000000  271.75183   1.000000 198.93300   1.000000
#> [45,]   7.267735   1.000000  877.03762   1.000000   1.00000  59.666673
#> [46,] 245.124993 148.175844    1.00000   1.000000   1.00000   1.000000
#> [47,]   1.000000   1.000000    1.00000  12.713010 199.14215   1.000000
#> [48,]  79.951947  83.327970    1.00000  11.524091   1.00000   1.000000
#> [49,] 136.178989   1.000000 2790.16865  25.455403   1.00000 139.800564
#> [50,]  25.153191   1.000000 2224.67078  38.601567  92.57719 220.132741
#>               l          m          n         o         p         q          r
#>  [1,] 100.00000 100.000000 100.000000 100.00000 100.00000 100.00000 100.000000
#>  [2,]   1.00000   1.000000 211.535564   1.00000   1.00000   1.00000  78.030058
#>  [3,]   1.00000   1.000000  95.176658  33.24628 240.03429 180.77792   1.000000
#>  [4,]   1.00000  74.842445   1.000000  13.90755 323.60093   1.00000   1.000000
#>  [5,]  73.59184   1.000000 187.403686   1.00000  77.54349 412.76810   1.000000
#>  [6,]   1.00000  42.722325   1.000000   1.00000  40.44665   1.00000  61.587491
#>  [7,]   1.00000   1.000000   1.000000  65.83557   1.00000   1.00000   1.000000
#>  [8,]   1.00000   1.000000 124.167989   1.00000 606.65644   1.00000   4.789239
#>  [9,]   1.00000   1.000000   1.000000   1.00000  77.55500  26.09230   1.000000
#> [10,]   1.00000   1.000000   1.000000   1.00000   1.00000 288.76321   1.000000
#> [11,]   1.00000   1.000000   1.000000  35.08710   1.00000 162.05986   1.000000
#> [12,]   1.00000   1.000000  29.228604  40.07044   1.00000  25.33642 116.879227
#> [13,] 157.12944   1.000000  12.858327  40.07772   1.00000   1.00000   1.000000
#> [14,]   1.00000   1.464792  11.907698   1.00000  57.90168   1.00000 175.520404
#> [15,] 100.08685 116.019894   1.000000  23.96872  77.71131   1.00000   1.000000
#> [16,]   1.00000   1.000000   1.000000  62.91213 230.57681 454.05911   1.000000
#> [17,]   1.00000   1.000000   1.000000  82.97930   1.00000   1.00000   1.000000
#> [18,]   1.00000   8.013669  39.458014   1.00000 340.90530 201.15958 189.835159
#> [19,]   1.00000   1.000000   1.000000   1.00000   1.00000   1.00000   1.000000
#> [20,] 144.00293   1.000000   1.000000  96.71143   1.00000 296.29871   1.000000
#> [21,]   1.00000  26.720721 143.043938   1.00000   1.00000   1.00000  40.825370
#> [22,]  29.64494  15.171247 237.333035   1.00000 208.34458  25.54358  20.919788
#> [23,]  78.81067   1.000000   1.000000  29.73564   1.00000   1.00000   3.602871
#> [24,]  16.59108  28.383926 210.006921   1.00000 221.68253 160.69161   1.000000
#> [25,]   1.00000   1.000000   1.000000  18.17618   1.00000 263.89217   1.000000
#> [26,]  87.63083 101.196948   1.000000 122.06365   1.00000 346.93834   1.000000
#> [27,]  15.88325  57.771318   1.000000   1.00000   1.00000 141.85434   9.415862
#> [28,]   1.00000  68.493797   1.000000   1.00000   1.00000 327.82737   1.000000
#> [29,]   1.00000  44.035163 100.636442  30.86989   1.00000 221.67816   1.000000
#> [30,] 194.29177  35.764913   1.000000   1.00000   1.00000   1.00000   1.000000
#> [31,]   1.00000   1.000000   1.000000  52.87210 295.92241  84.84171   1.000000
#> [32,]   1.00000   1.000000   1.000000  58.78950   1.00000   1.00000 132.356214
#> [33,]  49.65114  37.522918 266.904682   1.00000 221.58203   1.00000   1.000000
#> [34,]  28.88517  51.423088   1.000000   1.00000 425.64173 217.15003  21.824579
#> [35,]   1.00000  53.736721   7.697676   1.00000   1.00000  96.26811   1.089210
#> [36,]   1.00000  21.387037   1.000000   1.00000  79.67785  81.34161  56.740913
#> [37,]   1.00000  12.330590   1.000000  51.83079   1.00000   1.00000  67.725673
#> [38,] 106.08783   1.000000  13.277058   1.00000   1.00000   1.00000  71.074210
#> [39,]  59.22875  16.575912  10.335241   1.00000  15.77813   1.00000  21.151910
#> [40,]   1.00000 107.221265   1.000000   1.00000  83.94636  70.30390   1.000000
#> [41,]   1.00000   4.716743   1.000000  25.63619   1.00000   1.00000   1.000000
#> [42,]  25.23870   1.000000   1.000000   1.00000 753.57410  99.95981  35.496508
#> [43,]  14.65098   1.000000   1.000000  57.63986   1.00000   1.00000 166.245232
#> [44,]  58.63171  40.699577   1.000000  21.66949  18.03159 325.20659  83.282874
#> [45,]   1.00000  79.843837   1.000000   1.00000 182.38065  81.91679  45.138954
#> [46,]  83.73778   1.000000 303.567028  53.22491   1.00000   1.00000   1.000000
#> [47,]   1.00000   8.378723   1.000000  25.24425   1.00000  42.87034  65.338528
#> [48,]   1.00000   5.027135  66.281812   1.00000 293.39167   1.00000   1.000000
#> [49,]  68.97636   1.000000   1.000000  15.02278 453.44396  66.00916  26.847819
#> [50,]  92.17508  83.262086   1.000000  42.08923   1.00000  43.31262 119.704911
#>                s          t         u          v          w         x
#>  [1,]  100.00000 100.000000 100.00000 100.000000  100.00000 100.00000
#>  [2,]  711.71250 104.434261   1.00000  87.404945 1681.30668   1.00000
#>  [3,]    1.00000   1.805610 141.95652   1.000000    1.00000 137.37591
#>  [4,]    1.00000  13.273278   1.00000   1.000000    1.00000   1.00000
#>  [5,]    1.00000 256.009845 349.36340  17.497700    1.00000 101.67617
#>  [6,] 1148.12267   1.000000 176.41659   1.000000    1.00000 176.92369
#>  [7,]    1.00000   1.000000   1.00000   1.000000  581.96007   1.00000
#>  [8,]    1.00000   1.000000  16.45974   9.162817    1.00000 178.75210
#>  [9,] 1006.01074   1.000000 216.13074   1.000000 2081.06116   1.00000
#> [10,]    1.00000  44.884503 168.27794  17.601283 1467.23929  47.48784
#> [11,]    1.00000   1.000000 447.12740   1.000000    1.00000  32.56231
#> [12,]    1.00000   1.000000   1.00000 104.494822  584.72491   1.00000
#> [13,]    1.00000 122.385356 336.75944 264.684516  115.78081  43.25286
#> [14,]    1.00000   1.000000 312.74606   1.000000  202.24790   1.00000
#> [15,]    1.00000  21.213817   1.00000 115.466522 1141.84048   1.00000
#> [16,]    1.00000   1.000000   1.00000  90.082629 3543.39081 133.22429
#> [17,] 1308.54437   1.000000   1.00000   1.000000    1.00000   1.00000
#> [18,] 1183.00033 126.722442 138.12295   1.000000    1.00000   1.00000
#> [19,]   54.89027   1.000000   1.00000 171.978315 2262.07506   1.00000
#> [20,]  411.34090   1.000000 536.78878   1.000000    1.00000   1.00000
#> [21,]    1.00000  16.478230 161.46120   1.000000    1.00000  67.69105
#> [22,] 1494.53639  16.361410   1.00000   1.000000    1.00000   1.00000
#> [23,] 1060.93700   1.000000   1.00000 152.328115 1680.11498   1.00000
#> [24,]  779.74638   1.000000   1.00000   1.000000    1.00000   1.00000
#> [25,]    1.00000 165.544982 103.45589   1.000000  175.55384   1.00000
#> [26,]  777.62731   1.000000   1.00000 347.606122    1.00000 138.22463
#> [27,]  565.98047   1.000000 155.15291   1.000000 1010.80248   1.00000
#> [28,]    1.00000   1.000000   1.00000   1.000000   17.72438 187.35002
#> [29,]  228.47398   1.000000   1.00000  32.736665 3122.06860  64.26759
#> [30,]    1.00000   1.000000   1.00000   1.000000 1364.95376  58.45803
#> [31,]    1.00000 120.198508 574.78753  53.526953    1.00000   1.00000
#> [32,]  796.11332   1.000000 100.35824   1.000000 1239.80143 315.18394
#> [33,]  216.24369  95.847576  69.58651  51.149474   76.47723   1.00000
#> [34,]  365.81695  45.995045   1.00000   1.000000  821.34736   1.00000
#> [35,]    1.00000   1.000000 199.07164   1.000000    1.00000   1.00000
#> [36,]  544.72275 114.168966   1.00000   1.000000    1.00000   1.00000
#> [37,]   58.10979   1.000000   1.00000   1.000000 1443.73127   1.00000
#> [38,]    1.00000   1.000000 277.08083 141.818912  538.93923   1.00000
#> [39,]  749.38225   1.000000   1.00000 172.397298    1.00000 109.99913
#> [40,]    1.00000   1.000000   1.00000   1.000000 3859.75540 162.86135
#> [41,]    1.00000 102.666061 534.01340   1.000000  561.40915   1.00000
#> [42,]  760.43696   1.000000   1.00000   1.000000    1.00000 130.04901
#> [43,]    1.00000   8.930577 172.82803   7.558633    1.00000 229.78510
#> [44,]  151.00134  26.193270 306.92326  10.579157    1.00000   1.00000
#> [45,]    1.00000   1.000000   1.00000  31.663417 1515.88706   1.00000
#> [46,]   63.69395 102.470566 134.04977   1.000000    1.00000 104.98964
#> [47,]    1.00000 131.899237 132.49391   1.000000 1145.81057  29.53101
#> [48,]  956.45163  49.656709 313.97560 130.686468    1.00000  88.53350
#> [49,]  369.96333   1.000000 146.06190   1.000000 1415.61365 226.61678
#> [50,]  176.40350   1.000000   1.00000  33.083878    1.00000  62.44786
#>                y         z  indicator
#>  [1,] 100.000000 100.00000 100.000000
#>  [2,]  59.228365  42.89649  18.434784
#>  [3,]  41.230376 153.09207  14.593365
#>  [4,]  36.710466   1.00000   7.030356
#>  [5,] 122.784230   1.00000  13.449395
#>  [6,]   1.000000   1.00000   8.085288
#>  [7,]   1.000000 162.70082   6.706895
#>  [8,]  56.284614   1.00000   5.384031
#>  [9,]   1.000000 125.92040  10.927643
#> [10,]   7.036714   1.00000   6.618135
#> [11,]  70.483644   1.00000   6.922303
#> [12,]  71.626332  67.16126   6.900736
#> [13,]   1.000000   1.00000   9.254198
#> [14,]   1.000000  99.99237   8.342027
#> [15,]   6.998053  63.82710   8.984141
#> [16,] 101.076020   1.00000  15.079417
#> [17,]  40.507747  90.58142   8.378396
#> [18,] 136.256136 227.14503  19.490218
#> [19,]   1.000000   1.00000   6.211307
#> [20,] 107.033018   1.00000  11.895386
#> [21,]  83.959422  77.84705   6.330338
#> [22,]  58.199365 265.97442   8.679146
#> [23,]  39.431416   1.00000   8.705079
#> [24,]   1.000000   1.00000   9.699392
#> [25,]   1.000000   1.00000   4.393866
#> [26,]  21.811849   1.00000  15.012011
#> [27,]   4.648553   1.00000   6.033059
#> [28,]  89.685588  93.60815   3.937974
#> [29,]   1.000000  46.84329  11.003278
#> [30,]   1.000000 230.36703   4.861715
#> [31,]   1.000000  98.24635   7.490526
#> [32,]   1.000000   1.00000   9.594441
#> [33,]   1.000000  38.77631  11.887064
#> [34,]   1.000000  28.57198  14.172596
#> [35,]   1.000000  15.01343   3.437549
#> [36,]   1.000000 248.57982   9.733159
#> [37,]   1.000000   1.00000   8.415488
#> [38,]   1.000000 192.53611   6.202011
#> [39,]   1.000000  27.81953  13.540560
#> [40,]   1.000000  58.49652   7.363468
#> [41,]  62.137014  29.15650   6.858212
#> [42,] 168.606683   1.00000   8.816510
#> [43,]   1.000000  77.63637  10.537778
#> [44,] 141.978039 104.45120  15.475271
#> [45,]   7.894082 202.69464  10.250864
#> [46,]  10.985259 107.46570  11.875874
#> [47,]  30.885701 155.54817  12.770525
#> [48,]   1.000000   1.00000  10.406502
#> [49,]   1.000000 115.10932  14.701704
#> [50,]  86.378872   1.00000  15.526217

# Now, I have 'messed up' the data a bit by including two species with data missing at the beginning and one species with data missing at the end. We also have one species with some very high values.
trends_summary[1:3, 'a'] <- NA
trends_summary[1:5, 'b'] <- NA
trends_summary[2:4, 'c'] <- 1000
trends_summary[45:50, 'd'] <- NA

# Let's run this data again through our scaling function (all defaults used)
rescale_species(Data = trends_summary)
#>              year         a          b            c          d         e
#>  [1,] -0.56047565        NA         NA   100.000000 100.000000 100.00000
#>  [2,] -0.23017749        NA         NA 10000.000000  59.687411 149.59320
#>  [3,]  1.55870831        NA         NA 10000.000000   1.000000  91.56406
#>  [4,]  0.07050839  7.942748         NA 10000.000000  24.703998   1.00000
#>  [5,]  0.12928774  1.000000         NA     1.000000   1.000000   1.00000
#>  [6,]  1.71506499  8.800908   6.896343     1.000000   1.000000  23.57946
#>  [7,]  0.46091621  1.000000 120.214163    71.469058   1.000000   1.00000
#>  [8,] -1.26506123  3.392833 255.458150     1.000000   1.000000   1.00000
#>  [9,] -0.68685285  1.000000  58.234619   124.022497  75.081849  30.25533
#> [10,] -0.44566197  1.253227   1.000000     1.000000   1.000000 408.11775
#> [11,]  1.22408180  2.203256  88.118817   133.637115   5.423171 138.74157
#> [12,]  0.35981383  1.000000   1.000000     1.000000  11.082694 130.42245
#> [13,]  0.40077145  1.000000 247.791193     1.000000  56.051941   1.00000
#> [14,]  0.11068272  1.000000   8.509743   411.435839   1.000000   1.00000
#> [15,] -0.55584113  1.000000   1.000000     1.000000   1.000000   1.00000
#> [16,]  1.78691314  1.761543   1.000000    37.858688  76.209253   1.00000
#> [17,]  0.49785048  2.601207   1.000000    80.809735   1.000000  35.44993
#> [18,] -1.96661716  1.000000  98.129058     1.000000   1.000000 467.65548
#> [19,]  0.70135590  5.352423 130.138763    65.613375   1.000000 103.50822
#> [20,] -0.47279141 11.897763 156.853209    46.838432   1.000000   1.00000
#> [21,] -1.06782371  1.000000   1.000000     1.000000   1.000000   1.00000
#> [22,] -0.21797491  1.000000 145.113032     8.288665  28.105462   1.00000
#> [23,] -1.02600445  5.836851  75.132649     1.000000  50.474937   1.00000
#> [24,] -0.72889123  1.000000  39.222491   270.197656  32.180509 371.47595
#> [25,] -0.62503927  1.000000   1.000000     1.000000   1.000000   1.00000
#> [26,] -1.68669331  5.951952  99.851147     1.000000   2.717376 118.89079
#> [27,]  0.83778704  1.000000   1.000000     4.797072   1.000000   1.00000
#> [28,]  0.15337312  1.000000   1.000000    39.414173   1.000000   1.00000
#> [29,] -1.13813694  1.052203 147.315749    55.414746  40.233142   1.00000
#> [30,]  1.25381492  1.000000  10.921382     1.000000   1.000000   1.00000
#> [31,]  0.42646422  1.000000   1.000000     1.000000  88.925085 443.94631
#> [32,] -0.29507148  2.235993   1.000000   160.355831   1.000000   1.00000
#> [33,]  0.89512566  1.000000   1.000000     1.000000   9.757041   1.00000
#> [34,]  0.87813349  3.739670  64.708643     1.000000   1.000000  70.72673
#> [35,]  0.82158108  1.000000 314.470621     1.000000   1.000000   1.00000
#> [36,]  0.68864025  1.925512   1.000000     1.000000   1.000000   1.00000
#> [37,]  0.55391765  6.365557 223.708273   140.899524   1.000000   1.00000
#> [38,] -0.06191171  2.525596   1.000000    10.757028  19.054959   1.00000
#> [39,] -0.30596266  1.000000   1.000000    95.723829  14.749082  58.31968
#> [40,] -0.38047100  6.667159 221.143355     1.000000   1.000000   1.00000
#> [41,] -0.69470698  5.765846   1.000000    27.222894   1.000000   1.00000
#> [42,] -0.20791728  3.182648  40.157564     1.000000   1.000000   1.00000
#> [43,] -1.26539635  1.385491 240.785983    12.006965  68.039550   1.00000
#> [44,]  2.16895597  1.000000 231.983013     1.000000   1.000000 153.74430
#> [45,]  1.20796200  7.896610 245.287596     1.000000   1.000000   1.00000
#> [46,] -1.12310858  1.000000  81.312422   253.537501   1.000000   1.00000
#> [47,] -0.40288484 12.694290 223.879122    76.257357   1.000000   1.00000
#> [48,] -0.46665535  8.894578   1.000000     1.000000   1.000000 359.66250
#> [49,]  0.77996512  1.000000   1.000000     1.000000   1.000000   1.00000
#> [50,] -0.08336907  1.000000 197.118627     1.000000   1.000000   1.00000
#>                f          g          h          i         j          k
#>  [1,] 100.000000 100.000000  100.00000 100.000000 100.00000 100.000000
#>  [2,] 105.235539   1.000000 1588.79102  73.170241 165.09560   1.000000
#>  [3,] 131.219707   1.000000  862.94537  30.431225   1.00000 131.460036
#>  [4,] 147.154810  11.492001   39.21032  49.998409   1.00000  29.568406
#>  [5,]  61.120491   1.000000    1.00000  64.120065 250.73675 195.515102
#>  [6,]   1.000000  32.898684 2243.93131   1.000000  15.80804   1.000000
#>  [7,] 281.612373  40.537237  475.49370  77.619916 148.85504   1.000000
#>  [8,]   1.000000   1.000000    1.00000   1.000000 344.03982   1.000000
#>  [9,]   1.000000   1.000000  732.51539  16.122515   1.00000   1.000000
#> [10,]   1.000000 253.359815    1.00000   1.000000  13.16043   1.000000
#> [11,]   1.000000   1.000000    1.00000  60.959419  16.17718   1.000000
#> [12,]   1.000000  64.160566    1.00000   1.000000   1.00000  68.495617
#> [13,] 241.418981  26.973578    1.00000  36.248806   1.00000   1.000000
#> [14,]  84.098326 100.958680  166.82341   1.000000   1.00000 175.644867
#> [15,]  49.220594  80.562091  562.39655   1.000000 102.41621   1.000000
#> [16,]   1.000000   1.000000 3593.38227  84.592319 122.08141  11.963385
#> [17,]  14.774203  37.259995  126.35406  51.796616  21.89805  63.082804
#> [18,] 175.975166   1.000000    1.00000 120.543863   1.00000  69.444004
#> [19,]   1.000000  84.430639    1.00000   4.554937 172.73512   1.000000
#> [20,]   1.000000   1.000000  754.90539  78.649392  30.62154   9.409586
#> [21,]   1.000000 238.119081    1.00000 138.102336   1.00000 113.771964
#> [22,]   1.000000   1.000000    1.00000   1.000000  17.98993   1.000000
#> [23,] 187.177758   1.000000    1.00000   1.000000 116.03738   1.000000
#> [24,]   1.000000  81.322766    1.00000   1.000000  45.84623  37.921731
#> [25,]  73.109834  50.262178  976.76061   1.000000   1.00000   1.000000
#> [26,]   1.000000   1.000000  612.23898  10.604045   1.00000 162.149026
#> [27,]   8.503687   1.000000    1.00000 119.707926   1.00000  79.684044
#> [28,]   1.000000  14.234857    1.00000   1.000000   1.00000   8.541412
#> [29,]   1.000000   1.000000    1.00000  26.076905   1.00000   1.000000
#> [30,]   1.000000   1.000000 1955.66136   1.000000  92.78344   3.089134
#> [31,]   1.000000   3.404237  259.00913   1.429717   1.00000   1.000000
#> [32,] 166.158269  18.742952    1.00000  21.955897  84.12354   1.000000
#> [33,] 100.889524  17.215387    1.00000  92.856021 236.01635   1.000000
#> [34,]   1.000000   1.000000 1367.17059   8.481416   1.00000  59.019475
#> [35,]   1.000000  46.912309    1.00000  49.835092   1.00000   1.000000
#> [36,] 286.926982 135.827323 1464.76205  54.450416   1.00000   1.000000
#> [37,] 190.709871  44.951916    1.00000  63.952165   1.00000  45.144267
#> [38,]  28.071752   1.000000    1.00000   1.000000   1.00000   1.000000
#> [39,]   1.000000   1.000000 3024.88868 113.735917 101.10723  69.000885
#> [40,]  14.244581  34.100788    1.00000   1.000000  33.60081   1.000000
#> [41,]   1.000000   1.000000 1687.20998   1.000000  45.39813   1.000000
#> [42,]   1.000000   1.000000    1.00000  98.157716  77.87097  55.032034
#> [43,]   1.000000  87.123185    1.00000  90.469924   1.00000  99.483443
#> [44,]   1.000000   1.000000  271.75183   1.000000 198.93300   1.000000
#> [45,]   7.267735   1.000000  877.03762   1.000000   1.00000  59.666673
#> [46,] 245.124993 148.175844    1.00000   1.000000   1.00000   1.000000
#> [47,]   1.000000   1.000000    1.00000  12.713010 199.14215   1.000000
#> [48,]  79.951947  83.327970    1.00000  11.524091   1.00000   1.000000
#> [49,] 136.178989   1.000000 2790.16865  25.455403   1.00000 139.800564
#> [50,]  25.153191   1.000000 2224.67078  38.601567  92.57719 220.132741
#>               l          m          n         o         p         q          r
#>  [1,] 100.00000 100.000000 100.000000 100.00000 100.00000 100.00000 100.000000
#>  [2,]   1.00000   1.000000 211.535564   1.00000   1.00000   1.00000  78.030058
#>  [3,]   1.00000   1.000000  95.176658  33.24628 240.03429 180.77792   1.000000
#>  [4,]   1.00000  74.842445   1.000000  13.90755 323.60093   1.00000   1.000000
#>  [5,]  73.59184   1.000000 187.403686   1.00000  77.54349 412.76810   1.000000
#>  [6,]   1.00000  42.722325   1.000000   1.00000  40.44665   1.00000  61.587491
#>  [7,]   1.00000   1.000000   1.000000  65.83557   1.00000   1.00000   1.000000
#>  [8,]   1.00000   1.000000 124.167989   1.00000 606.65644   1.00000   4.789239
#>  [9,]   1.00000   1.000000   1.000000   1.00000  77.55500  26.09230   1.000000
#> [10,]   1.00000   1.000000   1.000000   1.00000   1.00000 288.76321   1.000000
#> [11,]   1.00000   1.000000   1.000000  35.08710   1.00000 162.05986   1.000000
#> [12,]   1.00000   1.000000  29.228604  40.07044   1.00000  25.33642 116.879227
#> [13,] 157.12944   1.000000  12.858327  40.07772   1.00000   1.00000   1.000000
#> [14,]   1.00000   1.464792  11.907698   1.00000  57.90168   1.00000 175.520404
#> [15,] 100.08685 116.019894   1.000000  23.96872  77.71131   1.00000   1.000000
#> [16,]   1.00000   1.000000   1.000000  62.91213 230.57681 454.05911   1.000000
#> [17,]   1.00000   1.000000   1.000000  82.97930   1.00000   1.00000   1.000000
#> [18,]   1.00000   8.013669  39.458014   1.00000 340.90530 201.15958 189.835159
#> [19,]   1.00000   1.000000   1.000000   1.00000   1.00000   1.00000   1.000000
#> [20,] 144.00293   1.000000   1.000000  96.71143   1.00000 296.29871   1.000000
#> [21,]   1.00000  26.720721 143.043938   1.00000   1.00000   1.00000  40.825370
#> [22,]  29.64494  15.171247 237.333035   1.00000 208.34458  25.54358  20.919788
#> [23,]  78.81067   1.000000   1.000000  29.73564   1.00000   1.00000   3.602871
#> [24,]  16.59108  28.383926 210.006921   1.00000 221.68253 160.69161   1.000000
#> [25,]   1.00000   1.000000   1.000000  18.17618   1.00000 263.89217   1.000000
#> [26,]  87.63083 101.196948   1.000000 122.06365   1.00000 346.93834   1.000000
#> [27,]  15.88325  57.771318   1.000000   1.00000   1.00000 141.85434   9.415862
#> [28,]   1.00000  68.493797   1.000000   1.00000   1.00000 327.82737   1.000000
#> [29,]   1.00000  44.035163 100.636442  30.86989   1.00000 221.67816   1.000000
#> [30,] 194.29177  35.764913   1.000000   1.00000   1.00000   1.00000   1.000000
#> [31,]   1.00000   1.000000   1.000000  52.87210 295.92241  84.84171   1.000000
#> [32,]   1.00000   1.000000   1.000000  58.78950   1.00000   1.00000 132.356214
#> [33,]  49.65114  37.522918 266.904682   1.00000 221.58203   1.00000   1.000000
#> [34,]  28.88517  51.423088   1.000000   1.00000 425.64173 217.15003  21.824579
#> [35,]   1.00000  53.736721   7.697676   1.00000   1.00000  96.26811   1.089210
#> [36,]   1.00000  21.387037   1.000000   1.00000  79.67785  81.34161  56.740913
#> [37,]   1.00000  12.330590   1.000000  51.83079   1.00000   1.00000  67.725673
#> [38,] 106.08783   1.000000  13.277058   1.00000   1.00000   1.00000  71.074210
#> [39,]  59.22875  16.575912  10.335241   1.00000  15.77813   1.00000  21.151910
#> [40,]   1.00000 107.221265   1.000000   1.00000  83.94636  70.30390   1.000000
#> [41,]   1.00000   4.716743   1.000000  25.63619   1.00000   1.00000   1.000000
#> [42,]  25.23870   1.000000   1.000000   1.00000 753.57410  99.95981  35.496508
#> [43,]  14.65098   1.000000   1.000000  57.63986   1.00000   1.00000 166.245232
#> [44,]  58.63171  40.699577   1.000000  21.66949  18.03159 325.20659  83.282874
#> [45,]   1.00000  79.843837   1.000000   1.00000 182.38065  81.91679  45.138954
#> [46,]  83.73778   1.000000 303.567028  53.22491   1.00000   1.00000   1.000000
#> [47,]   1.00000   8.378723   1.000000  25.24425   1.00000  42.87034  65.338528
#> [48,]   1.00000   5.027135  66.281812   1.00000 293.39167   1.00000   1.000000
#> [49,]  68.97636   1.000000   1.000000  15.02278 453.44396  66.00916  26.847819
#> [50,]  92.17508  83.262086   1.000000  42.08923   1.00000  43.31262 119.704911
#>                s          t         u          v          w         x
#>  [1,]  100.00000 100.000000 100.00000 100.000000  100.00000 100.00000
#>  [2,]  711.71250 104.434261   1.00000  87.404945 1681.30668   1.00000
#>  [3,]    1.00000   1.805610 141.95652   1.000000    1.00000 137.37591
#>  [4,]    1.00000  13.273278   1.00000   1.000000    1.00000   1.00000
#>  [5,]    1.00000 256.009845 349.36340  17.497700    1.00000 101.67617
#>  [6,] 1148.12267   1.000000 176.41659   1.000000    1.00000 176.92369
#>  [7,]    1.00000   1.000000   1.00000   1.000000  581.96007   1.00000
#>  [8,]    1.00000   1.000000  16.45974   9.162817    1.00000 178.75210
#>  [9,] 1006.01074   1.000000 216.13074   1.000000 2081.06116   1.00000
#> [10,]    1.00000  44.884503 168.27794  17.601283 1467.23929  47.48784
#> [11,]    1.00000   1.000000 447.12740   1.000000    1.00000  32.56231
#> [12,]    1.00000   1.000000   1.00000 104.494822  584.72491   1.00000
#> [13,]    1.00000 122.385356 336.75944 264.684516  115.78081  43.25286
#> [14,]    1.00000   1.000000 312.74606   1.000000  202.24790   1.00000
#> [15,]    1.00000  21.213817   1.00000 115.466522 1141.84048   1.00000
#> [16,]    1.00000   1.000000   1.00000  90.082629 3543.39081 133.22429
#> [17,] 1308.54437   1.000000   1.00000   1.000000    1.00000   1.00000
#> [18,] 1183.00033 126.722442 138.12295   1.000000    1.00000   1.00000
#> [19,]   54.89027   1.000000   1.00000 171.978315 2262.07506   1.00000
#> [20,]  411.34090   1.000000 536.78878   1.000000    1.00000   1.00000
#> [21,]    1.00000  16.478230 161.46120   1.000000    1.00000  67.69105
#> [22,] 1494.53639  16.361410   1.00000   1.000000    1.00000   1.00000
#> [23,] 1060.93700   1.000000   1.00000 152.328115 1680.11498   1.00000
#> [24,]  779.74638   1.000000   1.00000   1.000000    1.00000   1.00000
#> [25,]    1.00000 165.544982 103.45589   1.000000  175.55384   1.00000
#> [26,]  777.62731   1.000000   1.00000 347.606122    1.00000 138.22463
#> [27,]  565.98047   1.000000 155.15291   1.000000 1010.80248   1.00000
#> [28,]    1.00000   1.000000   1.00000   1.000000   17.72438 187.35002
#> [29,]  228.47398   1.000000   1.00000  32.736665 3122.06860  64.26759
#> [30,]    1.00000   1.000000   1.00000   1.000000 1364.95376  58.45803
#> [31,]    1.00000 120.198508 574.78753  53.526953    1.00000   1.00000
#> [32,]  796.11332   1.000000 100.35824   1.000000 1239.80143 315.18394
#> [33,]  216.24369  95.847576  69.58651  51.149474   76.47723   1.00000
#> [34,]  365.81695  45.995045   1.00000   1.000000  821.34736   1.00000
#> [35,]    1.00000   1.000000 199.07164   1.000000    1.00000   1.00000
#> [36,]  544.72275 114.168966   1.00000   1.000000    1.00000   1.00000
#> [37,]   58.10979   1.000000   1.00000   1.000000 1443.73127   1.00000
#> [38,]    1.00000   1.000000 277.08083 141.818912  538.93923   1.00000
#> [39,]  749.38225   1.000000   1.00000 172.397298    1.00000 109.99913
#> [40,]    1.00000   1.000000   1.00000   1.000000 3859.75540 162.86135
#> [41,]    1.00000 102.666061 534.01340   1.000000  561.40915   1.00000
#> [42,]  760.43696   1.000000   1.00000   1.000000    1.00000 130.04901
#> [43,]    1.00000   8.930577 172.82803   7.558633    1.00000 229.78510
#> [44,]  151.00134  26.193270 306.92326  10.579157    1.00000   1.00000
#> [45,]    1.00000   1.000000   1.00000  31.663417 1515.88706   1.00000
#> [46,]   63.69395 102.470566 134.04977   1.000000    1.00000 104.98964
#> [47,]    1.00000 131.899237 132.49391   1.000000 1145.81057  29.53101
#> [48,]  956.45163  49.656709 313.97560 130.686468    1.00000  88.53350
#> [49,]  369.96333   1.000000 146.06190   1.000000 1415.61365 226.61678
#> [50,]  176.40350   1.000000   1.00000  33.083878    1.00000  62.44786
#>                y         z  indicator
#>  [1,] 100.000000 100.00000 100.000000
#>  [2,]  59.228365  42.89649  28.502036
#>  [3,]  41.230376 153.09207  19.766610
#>  [4,]  36.710466   1.00000   7.942748
#>  [5,] 122.784230   1.00000  12.267967
#>  [6,]   1.000000   1.00000   6.896343
#>  [7,]   1.000000 162.70082   6.728697
#>  [8,]  56.284614   1.00000   4.592307
#>  [9,]   1.000000 125.92040   9.439848
#> [10,]   7.036714   1.00000   5.626644
#> [11,]  70.483644   1.00000   5.904375
#> [12,]  71.626332  67.16126   6.900736
#> [13,]   1.000000   1.00000   9.284280
#> [14,]   1.000000  99.99237   8.369144
#> [15,]   6.998053  63.82710   8.984141
#> [16,] 101.076020   1.00000  12.820306
#> [17,]  40.507747  90.58142   7.123193
#> [18,] 136.256136 227.14503  17.395310
#> [19,]   1.000000   1.00000   5.297931
#> [20,] 107.033018   1.00000  10.146164
#> [21,]  83.959422  77.84705   6.330338
#> [22,]  58.199365 265.97442   8.707359
#> [23,]  39.431416   1.00000   7.424993
#> [24,]   1.000000   1.00000   9.730921
#> [25,]   1.000000   1.00000   4.393866
#> [26,]  21.811849   1.00000  12.804487
#> [27,]   4.648553   1.00000   6.033059
#> [28,]  89.685588  93.60815   3.937974
#> [29,]   1.000000  46.84329   9.385240
#> [30,]   1.000000 230.36703   4.877519
#> [31,]   1.000000  98.24635   7.257362
#> [32,]   1.000000   1.00000   8.157058
#> [33,]   1.000000  38.77631  11.887064
#> [34,]   1.000000  28.57198  12.088508
#> [35,]   1.000000  15.01343   3.448723
#> [36,]   1.000000 248.57982   8.274994
#> [37,]   1.000000   1.00000   7.177986
#> [38,]   1.000000 192.53611   5.272862
#> [39,]   1.000000  27.81953  13.540560
#> [40,]   1.000000  58.49652   6.280666
#> [41,]  62.137014  29.15650   5.830755
#> [42,] 168.606683   1.00000   7.520038
#> [43,]   1.000000  77.63637   8.988192
#> [44,] 141.978039 104.45120  15.525576
#> [45,]   7.894082 202.69464   8.743469
#> [46,]  10.985259 107.46570  10.036238
#> [47,]  30.885701 155.54817  10.892613
#> [48,]   1.000000   1.00000   8.847461
#> [49,]   1.000000 115.10932  14.701704
#> [50,]  86.378872   1.00000  13.828788